Add eratosthenes sieve lib in C, and also solve problem number 10 in C

+++ Active/C/tester.h
@@ -3,11 +3,12 @@
 #include <stdio.h>
 #include <stdbool.h>
 
-void test(char str[], unsigned long int result, unsigned long int answer) {
-    if (result ==  answer) {
+void test(char str[], unsigned long long int result, unsigned long long int answer) {
+    if (result == answer) {
         printf("Solution for problem %s is correct.\n", str);
-    } else {
+    }
+    else {
         printf("Solution for problem %s is INCORRECT.\n", str);
-        printf("\tExpected: %lu\n\tReceived: %lu\n", answer, result);
+        printf("\tExpected: %llu\n\tReceived: %llu\n", answer, result);
     }
 }
